move resource modules into separate files
