+++ lib/protip/resource.rb
@@ -24,111 +24,18 @@ require 'protip/wrapper'
 
 require 'protip/messages/array'
 
+require 'protip/resource/creatable'
+require 'protip/resource/updateable'
+require 'protip/resource/destroyable'
+require 'protip/resource/extra_methods'
+require 'protip/resource/search_methods'
+
 module Protip
   module Resource
-
-    # Internal handlers for index/show actions. Never use these directly; instead, use `.all` and
-    # `.find` on the resource you're working with, since those methods will adjust their
-    # signatures to correctly parse a set of query parameters if supported.
-    module SearchMethods
-      # Fetch a list from the server at the collection's base endpoint. Expects the server response
-      # to be an array containing encoded messages that can be used to instantiate our resource.
-      #
-      # @param resource_class [Class] The resource type that we're fetching.
-      # @param query [::Protobuf::Message|NilClass] An optional query to send along with the request.
-      # @return [Array] The array of resources (each is an instance of the resource class we were
-      #   initialized with).
-      def self.index(resource_class, query)
-        response = resource_class.client.request path: resource_class.base_path,
-          method: Net::HTTP::Get,
-          message: query,
-          response_type: Protip::Messages::Array
-        response.messages.map do |message|
-          resource_class.new resource_class.message.decode(message)
-        end
-      end
-
-      # Fetch a single resource from the server.
-      #
-      # @param resource_class [Class] The resource type that we're fetching.
-      # @param id [String] The ID to be used in the URL to fetch the resource.
-      # @param query [::Protobuf::Message|NilClass] An optional query to send along with the request.
-      # @return [Protip::Resource] An instance of our resource class, created from the server
-      #   response.
-      def self.show(resource_class, id, query)
-        response = resource_class.client.request path: "#{resource_class.base_path}/#{id}",
-          method: Net::HTTP::Get,
-          message: query,
-          response_type: resource_class.message
-        resource_class.new response
-      end
-    end
-
-    # Mixin for a resource that has an active `:create` action. Should be treated as private,
-    # and will be included automatically when appropriate.
-    module Creatable
-      private
-      # POST the resource to the server and update our internal message. Private, since
-      # we should generally do this through the `save` method.
-      def create!
-        raise RuntimeError.new("Can't re-create a persisted object") if persisted?
-        self.message = self.class.client.request path: self.class.base_path,
-          method: Net::HTTP::Post,
-          message: message,
-          response_type: self.class.message
-      end
-    end
-
-    # Mixin for a resource that has an active `:update` action. Should be treated as private,
-    # and will be included automatically when appropriate.
-    module Updatable
-      private
-      # PUT the resource on the server and update our internal message. Private, since
-      # we should generally do this through the `save` method.
-      def update!
-        raise RuntimeError.new("Can't update a non-persisted object") if !persisted?
-        self.message = self.class.client.request path: "#{self.class.base_path}/#{id}",
-          method: Net::HTTP::Put,
-          message: message,
-          response_type: self.class.message
-      end
-    end
-
-    # Mixin for a resource that has an active `:destroy` action. Should be treated as private,
-    # and will be included automatically when appropriate.
-    module Destroyable
-      def destroy
-        raise RuntimeError.new("Can't destroy a non-persisted object") if !persisted?
-        self.message = self.class.client.request path: "#{self.class.base_path}/#{id}",
-          method: Net::HTTP::Delete,
-          message: nil,
-          response_type: self.class.message
-      end
-    end
-
-    # Internal helpers for non-resourceful member/collection methods. Never use these directly;
-    # instead, use the instance/class methods which have been dynamically defined on the resource
-    # you're working with.
-    module ExtraMethods
-      def self.member(resource, action, method, message, response_type)
-        response = resource.class.client.request path: "#{resource.class.base_path}/#{resource.id}/#{action}",
-          method: method,
-          message: message,
-          response_type: response_type
-        nil == response ? nil : ::Protip::Wrapper.new(response, resource.class.converter)
-      end
-      def self.collection(resource_class, action, method, message, response_type)
-        response = resource_class.client.request path: "#{resource_class.base_path}/#{action}",
-          method: method,
-          message: message,
-          response_type: response_type
-        nil == response ? nil : ::Protip::Wrapper.new(response, resource_class.converter)
-      end
-    end
-
     extend ActiveSupport::Concern
 
-    # Backport the ActiveModel::Model functionality - https://github.com/rails/rails/blob/097ca3f1f84bb9a2d3cda3f2cce7974a874efdf4/activemodel/lib/active_model/model.rb#L95
+    # Backport the ActiveModel::Model functionality
+    # https://github.com/rails/rails/blob/097ca3f1f84bb9a2d3cda3f2cce7974a874efdf4/activemodel/lib/active_model/model.rb#L95
     include ActiveModel::Validations
     include ActiveModel::Conversion
 
@@ -142,6 +49,7 @@ module Protip
       def_delegator :@wrapper, :message
       def_delegator :@wrapper, :as_json
     end
+
     module ClassMethods
 
       attr_accessor :client
@@ -150,7 +58,11 @@ module Protip
 
       attr_writer :base_path
       def base_path
-        @base_path == nil ? raise(RuntimeError.new 'Base path not yet set') : @base_path.gsub(/\/$/, '')
+        if @base_path == nil
+          raise(RuntimeError.new 'Base path not yet set')
+        else
+          @base_path.gsub(/\/$/, '')
+        end
       end
 
       attr_writer :converter
@@ -333,5 +245,6 @@ module Protip
       @previously_changed = changes
       @changed_attributes.clear
     end
+
   end
 end
